Add notes & buttons row to posts
The very bottom row of every post has the note count and 2-4 buttons
depending on what sort of post it is. Create very simple mockup buttons
(the sketchiness may actually be a good guide aesthetic wise - make the
app flow, like a painting). This technically also adds the tags line as a
TextView, but since we ultimately want the tags to create links we may
want to change that.