WIP Photo posts
This is the point where I realize that the sizes of the window suddenly
changing when I scroll up and down the page is jaring if you're looking at
the page position indicator. Thus I'm thinking Tumblr doesn't use this
view. Will look at using a simpler set of views that just load the
pictures when you first open the app.