WIP Post Source line
Getting an exception when I'm trying to load the image, but at least each
post has it's blog source.

EDIT: the current library calls this.requestBuilder.getRedirectUrl(),
which then triggers an error since we get the response code 200 (aka
HTTP_OK) as opposed to 301 (which indicates that it has a header which we
use to get the actual blog avatar URL). Not sure how to fix this. This is
the first time I've really tried this, so I'll post the bug or find the
original creators and see how things go.