Separate all posts into the originals and reblogged posts by blog & sorted chronologically
I realized that I had enough of the post functionality (the ability to
like posts, and seeing most of the content that I wanted) to shift gears
from emulating tumblr to using my own version. For this I wanted to be
able to 1. Have originals and reblogs presorted and 2. Sort everything by
chronological order, instead of reverse chronological order as tumblr
does.

To sort & categorize the posts in advance, I grab the posts via the
CompilerService, which runs every 15 minutes. This saves the post ids into
the shared preferences, which is documented to be good to use for an
activity which wants to store a small number of key value pairs. The
MainActivity counts and lists the categories, while the PostListActivity
shows the actual posts.

CompilerService is based on https://guides.codepath.com/android/Starting-Background-Services#using-with-alarmmanager-for-periodic-tasks