Add avatar
Using the jumblr api calls on an LG3 Phone with Marshmellow caused
a crash, specifically when attempting to get the avatar link. This
was mentioned in https://github.com/tumblr/jumblr/issues/94, but
the crash was not reproduciable by the developers. After several
more attempts on my side to figure out what I was doing differently
I resorted to using the deeper level API to get the avatar.
After getting that to work, I refactored the AuthorView with this
new capability in mind.