Handle screen rotation in PostListActivity
Before all the posts would be lost since the activity was recreated,
and the entry in the SharedPreferences had been erased. Now all the
posts will at least stay on the screen. It's debatable whether or
not the posts should expand in width as dramatically as they do,
but that's functionality that's due further down the line.