Configurable Categories
Instead of forcing reblogs & originals, make the categories more nuanced.
We can have the list of primary posts (those that we want to look at) and
the list of secondaries. The primaries currently are defined arbitrarily
(some I want to see everything, some I want to see just originals, and
some I just want to see originals that have images or video). Eventually
sort out which categories via a selection of some form.